refact: split between env configs and keys info
